backend/requirements/base.txt
amqp==5.3.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.1
attrs==26.1.0
bcrypt==5.0.0
billiard==4.2.4
celery==5.6.3
cffi==2.0.0
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==6.0.4
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-debug-toolbar==6.3.0
django-environ==0.13.0
django-filter==25.2
django-jazzmin==3.0.4
django-redis==6.0.0
django-rest-passwordreset==1.5.0
django-structurator==1.3.0
django-timezone-field==7.2.1
Django-Verify-Email==3.0.3
django_celery_results==2.6.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
inflection==0.5.1
joblib
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
numpy
packaging==26.1
pandas
prompt_toolkit==3.0.52
psycopg2==2.9.11
pycparser==3.0
PyJWT==2.12.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
tzdata==2026.1
tzlocal==5.3.1
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.6.0
